resize method created for layouts
